backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
distlib==0.3.9
Django==4.2
django-cors-headers==4.6.0
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.16.1
inflection==0.5.1
mysqlclient==2.2.6
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
PyJWT==2.10.1
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.28.0
